Implement authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -4,7 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Question;
 use Illuminate\Http\Request;
-use App\Http\Requests\AskQuestionRequest;
+use App\Http\Requests\AskQuestionRequest; 
+use App\Providers\AuthServiceProvider;
 
 class QuestionsController extends Controller
 {
@@ -66,7 +67,11 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
-        return view('questions.edit', compact('question'));     
+        if (\Gate::denies('update-question', $question)) {
+             abort(403, "Accenss denied");
+        }
+        return view('questions.edit', compact('question'));   
+
     }
 
     /**
@@ -78,6 +83,10 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, "Accenss denied");
+       }
+
         $question->update($request->only('title', 'body'));
 
         return redirect(('/questions'))->with('success', "Your question has been updated");
@@ -91,6 +100,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (\Gate::denies('delete-question', $question)) {
+            abort(403, "Accenss denied");
+       }
         $question->delete();
 
         return redirect('/questions')->with('success', "Your question has been deleted.");
